implemented default for MockDataDir
