+++ pubky-homeserver/src/data_directory/mod.rs
@@ -1,15 +1,15 @@
 mod config_toml;
+mod persistent_data_dir;
+mod mock_data_dir;
 mod data_dir;
-mod data_dir_mock;
-mod data_dir_trait;
 mod domain;
 mod domain_port;
 mod signup_mode;
 
 pub use config_toml::{ConfigReadError, ConfigToml};
-pub use data_dir::PersistentDataDir;
-pub use data_dir_mock::MockDataDir;
-pub use data_dir_trait::DataDir;
+pub use persistent_data_dir::PersistentDataDir;
+pub use mock_data_dir::MockDataDir;
+pub use data_dir::DataDir;
 pub use domain::Domain;
 pub use domain_port::DomainPort;
 pub use signup_mode::SignupMode;
